rename header.tsx to HeaderIcons.tsx and sidebar.tsx to SidebarIcons.tsx
